Add PlayButton component
Rename CardMediaType to PlayButton and move it to the shared
components.

The PlayButton has an optional label property. Pass the duration
through the label option.

Add a getter for the duration in the vuex store, and map the getter in
the cards mixin.

https://phabricator.endlessm.com/T31935
