+++ packages/template-ui/src/components/CarouselCard.vue
@@ -5,7 +5,7 @@
         <b-card>
           <template>
             <div class="img" :style="backgroundStyle">
-              <CardMediaType :node="node" />
+              <PlayButton :node="node" :label="label" @click="goToContent(node)" />
             </div>
             <b-card-text>
               <CardBody :node="node" :titleLines="5" />
@@ -19,6 +19,7 @@
 
 <script>
 import cardMixin from '@/components/mixins/cardMixin';
+import { goToContent } from 'kolibri-api';
 
 export default {
   name: 'CarouselCard',
@@ -32,6 +33,9 @@ export default {
         backgroundImage: `url("${this.thumbnail}")`,
       };
     },
+  },
+  methods: {
+    goToContent,
   },
 };
 </script>
